+++ CHANGELOG.md
@@ -32,6 +32,8 @@
     // logs == false
     const daoFundLsigProg = getProgram('dao-fund-lsig.py', {}, false);
     ```
++ Updated `deployer.deployApp()` & `deployer.updateApp()` to take one more optional parameter: `appName`. This will also save in a checkpoint the compiled app by name. Added new funtion `getAppByName(name: string)` to query checkpoint information by app name.
++ Added `deployer.loadLogicFromCache` to load a logic signature from already compiled TEAL codes (stored in `artifacts/cache`, for eg during `deployer.fundLsig`). This avoid re-compilation (and passing `scTmplParams`) each time(s) user wants to load an lsig.
 + Updated `TealDbg` method to load already compiled TEAL code from `artifacts/cache`. Compilation is forced only when a) TEAL is not cached OR b) `scInitParam` (template parameters) are passed with `tealFile`.
 
 ### Breaking changes